+++ commands/open.ts
@@ -1,0 +1,6 @@
+/**
+ * Запрашивает ресурс
+ */
+browser.addCommand('open', function (url) {
+	return browser.getUrl(url);
+});