+++ types/index.d.ts
@@ -2,7 +2,6 @@
 /// <reference path="./file-service.d.ts" />
 /// <reference path="./portal.d.ts" />
 /// <reference path="./process.d.ts" />
-/// <reference path="./tir.d.ts" />
 /// <reference path="./webdriver.d.ts" />
 /// <reference path="./wdio-utils.d.ts" />
 /// <reference path="./yoda.d.ts" />